consolidate buffer related operations
