+++ skyrim-extractor/src/binary-decoders/buffer.constants.ts
