+++ skyrim-extractor/src/binary-decoders/bufferParser.ts
